+++ lib/auto_html/auto_html_for.rb
@@ -29,7 +29,10 @@ module AutoHtmlFor
           write_attribute("#{raw_attr}#{suffix}", val)
         end
         define_method("#{raw_attr}#{suffix}") do
-          read_attribute("#{raw_attr}#{suffix}") || send("auto_html_prepare_#{raw_attr}")
+          result = read_attribute("#{raw_attr}#{suffix}") || send("auto_html_prepare_#{raw_attr}")
+          result.respond_to?(:html_safe) ?
+            result.html_safe :
+              result
         end
         define_method("auto_html_prepare_#{raw_attr}") do
           self.send(raw_attr.to_s + suffix + "=",